+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   - This no longer supports AuthorityImminentLoss.
 - ComponentUpdateSystem no longer has the API `GetAuthority`, `GetComponent`, and `HasComponent`. [#1364](https://github.com/spatialos/gdk-for-unity/pull/1364)
   - Use the Unity Entities `EntityManager` instead.
+- The GDK now depends on .NET Core v3.1.3xx instead of v2.2.2xx. [#1443](https://github.com/spatialos/gdk-for-unity/pull/1443)
+- Removed APIs for AuthorityLossImminent. [#1451](https://github.com/spatialos/gdk-for-unity/pull/1451)
+  - All authority changes to `AuthorityLossImminent` will now be dropped, and callbacks will no longer trigger.
 
 ### Added